// titmas-typescript-sdk/generated/src/models/PreflightRequest.ts
export interface PreflightRequest {
    /**
     *
     * @type {string}
     * @memberof PreflightRequest
     */
    agentId: string;
    /**
     *
     * @type {PreflightRequestApiVersionEnum}
     * @memberof PreflightRequest
     */
    apiVersion?: PreflightRequestApiVersionEnum;
    /**
     *
     * @type {string}
     * @memberof PreflightRequest
     */
    credentialId: string;
    /**
     *
     * @type {string}
     * @memberof PreflightRequest
     */
    idempotencyKey: string;
    /**
     *
     * @type {any}
     * @memberof PreflightRequest
     */
    object: any | null;
    /**
     *
     * @type {string}
     * @memberof PreflightRequest
     */
    objectDigest: string;
    /**
     *
     * @type {string}
     * @memberof PreflightRequest
     */
    requestId: string;
    /**
     *
     * @type {string}
     * @memberof PreflightRequest
     */
    schemaId: string;
    /**
     *
     * @type {string}
     * @memberof PreflightRequest
     */
    schemaVersion: string;
    /**
     *
     * @type {string}
     * @memberof PreflightRequest
     */
    tenantId: string;
    /**
     *
     * @type {Date}
     * @memberof PreflightRequest
     */
    timestamp?: Date;
}


/**
 * @export
 */
export const PreflightRequestApiVersionEnum = {
    V1: 'v1'
} as const;
export type PreflightRequestApiVersionEnum = typeof PreflightRequestApiVersionEnum[keyof typeof PreflightRequestApiVersionEnum];