// titmas-typescript-sdk/generated/src/models/ReceiptResponse.ts
export interface ReceiptResponse {
    /**
     *
     * @type {Array<ReasonCode>}
     * @memberof ReceiptResponse
     */
    reasonCodes: Array<ReasonCode>;
    /**
     *
     * @type {Receipt}
     * @memberof ReceiptResponse
     */
    receipt: Receipt;
    /**
     *
     * @type {string}
     * @memberof ReceiptResponse
     */
    receiptId: string;
    /**
     *
     * @type {string}
     * @memberof ReceiptResponse
     */
    requestDigest: string;
    /**
     *
     * @type {string}
     * @memberof ReceiptResponse
     */
    responseDigest: string;
    /**
     *
     * @type {Result}
     * @memberof ReceiptResponse
     */
    result: Result;
    /**
     *
     * @type {string}
     * @memberof ReceiptResponse
     */
    schemaId: string | null;
    /**
     *
     * @type {string}
     * @memberof ReceiptResponse
     */
    schemaVersion: string | null;
    /**
     *
     * @type {ReceiptResponseSignatureStatusEnum}
     * @memberof ReceiptResponse
     */
    signatureStatus: ReceiptResponseSignatureStatusEnum;
    /**
     *
     * @type {ReceiptVerificationStatus}
     * @memberof ReceiptResponse
     */
    verificationStatus: ReceiptVerificationStatus;
}


/**
 * @export
 */
export const ReceiptResponseSignatureStatusEnum = {
    Signed: 'SIGNED'
} as const;
export type ReceiptResponseSignatureStatusEnum = typeof ReceiptResponseSignatureStatusEnum[keyof typeof ReceiptResponseSignatureStatusEnum];